Version 2.0
It contains Default, Low Vision and High Contrast themes.
It contains all the images created by our graphical designer according to the wireframes.
The app is similar to our version 1 implementation and is used only for testing and validation puproses, since we don't have the source code of the app.